5/4 20:00 视觉方面：完成除了ScreenManager和SpiralField和主函数之外的主要内容

+++ ConsoleApplication1/CircularField.h
@@ -5,7 +5,7 @@
 class CircularField : public Field 
 {
 public:
-	CircularField(const double& creationTime,const Point& s_Position, const double& s_Amplitude, const double& s_Frequency, const double& s_InitialPhase, const int& s_Speed);
+	CircularField(const double& creationTime,const Point& s_Position, const double& s_Amplitude, const double& s_Frequency, const double& s_InitialPhase, const double& s_Speed);
 	virtual ~CircularField() {};
 	
 	void ActivateMedium (const double& timeOfNow, const double& frameTime) override;
